# InfluencerWebExperience/openapi.yaml
openapi: 3.0.3
info:
  title: InfluencerCRB Web Experience Service API
  version: 1.0.0
  description: Brand authentication API for signup, login, logout, and Google/Facebook social auth.
servers:
  - url: https://localhost:8081
paths:
...
  /api/import-batches:
    get:
      summary: List import batches
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Import batch list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "f2d37f81-0e40-4ff6-9ea2-a5a5985ec444"
                      userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                      status: "discovered"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "f2d37f81-0e40-4ff6-9ea2-a5a5985ec444"
                        userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                        status: "discovered"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
  /api/import-batches/{id}:
    get:
      summary: Get import batch by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Import batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/import-batches/discover:
    post:
      summary: Discover import schema from uploaded file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Discovery result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/import-batches/{id}/preview:
    post:
      summary: Preview import rows
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Preview result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
...
  /api/import-batches/{id}/hydrate:
    post:
      summary: Hydrate import data into domain records
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Hydration result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
...
  /api/influencer-campaign-codes:
    get:
      summary: List influencer campaign codes
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: creatorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Code list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "99a1418f-86e7-455b-a020-27f7d2532999"
                      campaignId: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                      creatorId: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                      code: "LENA10"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "99a1418f-86e7-455b-a020-27f7d2532999"
                        campaignId: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                        creatorId: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                        code: "LENA10"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create influencer campaign code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/influencer-campaign-codes/{id}:
    get:
      summary: Get influencer campaign code by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update influencer campaign code
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete influencer campaign code
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/influencer-sale-attributions:
    get:
      summary: List influencer sale attributions
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignCodeId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignCreatorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Attribution list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "deeb7df8-1e3c-4f99-a46a-cf9dbdbb0aaa"
                      campaignCodeId: "99a1418f-86e7-455b-a020-27f7d2532999"
                      orderId: "ORDER-1001"
                      status: "attributed"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "deeb7df8-1e3c-4f99-a46a-cf9dbdbb0aaa"
                        campaignCodeId: "99a1418f-86e7-455b-a020-27f7d2532999"
                        orderId: "ORDER-1001"
                        status: "attributed"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create influencer sale attribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created attribution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/influencer-sale-attributions/{id}:
    get:
      summary: Get influencer sale attribution by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Attribution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update influencer sale attribution
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated attribution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete influencer sale attribution
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      required: false
      description: 1-based page number. When omitted with size, defaults to 1.
      schema:
        type: integer
        minimum: 1
    SizeQuery:
      name: size
      in: query
      required: false
      description: Page size. When omitted with page, defaults to 20.
      schema:
        type: integer
        minimum: 1
  responses:
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GenericJson:
      type: object
      additionalProperties: true
    PaginatedGenericJson:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasPrevious:
          type: boolean
        hasNext:
          type: boolean
      required:
        - items
        - page
        - size
        - total
        - totalPages
        - hasPrevious
        - hasNext
    ListOrPaginatedGenericJson:
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        - $ref: '#/components/schemas/PaginatedGenericJson'
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - errorCode
        - message